# agent-manager-observer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Manager Observer API
  description: |
    Observer Service-backed API for listing, exporting, and fetching distributed traces,
    component/build logs, and resource metrics.

    Endpoints use human-readable `organization`, `project`, `agent`, and `environment` identifiers.
    Lazy loading is supported: the span list endpoint returns summaries only; individual span
    attributes are fetched on demand via the span detail endpoint.

    All endpoints require the same Bearer JWT as the Agent Manager Service console
    (validated via JWKS / local-dev mode).
  version: 1.0.0
  contact:
    name: WSO2 LLC
    url: https://www.wso2.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:9098
    description: Local development
  - url: https://observer.example.com
    description: Production (replace with your deployment URL)
...
paths:
...
  /api/v1/traces:
    get:
      tags:
        - traces
      summary: List traces
      description: |
        Returns paginated trace overviews for a component. Each trace includes input, output, token usage,
        and status extracted from the root span — the same richness as the v1 list endpoint.
        Uses the Observer Service instead of OpenSearch. No offset; use `limit` for page size.
      operationId: listTraces
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organization"
        - $ref: "#/components/parameters/project"
        - $ref: "#/components/parameters/agent"
        - $ref: "#/components/parameters/environment"
        - $ref: "#/components/parameters/startTimeRequired"
        - $ref: "#/components/parameters/endTimeRequired"
        - name: limit
          in: query
          required: false
          description: Page size (default 10, max 1000)
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 10
        - $ref: "#/components/parameters/sortOrder"
      responses:
        "200":
          description: Trace overview list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceOverviewListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ObserverUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/traces/export:
    get:
      tags:
        - traces
      summary: Export traces with full span details
      description: |
        Returns complete trace objects for a component including ALL spans, each fully enriched with
        AmpAttributes (LLM data, tool calls, token usage, etc.). Response includes Content-Disposition
        for download. Export limit is capped at 1000 traces.

        `organization`, `project`, `agent`, and `environment` are all required to scope the export
        to a specific agent.
      operationId: exportTraces
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organization"
        - $ref: "#/components/parameters/project"
        - $ref: "#/components/parameters/agent"
        - $ref: "#/components/parameters/environment"
        - $ref: "#/components/parameters/startTimeRequired"
        - $ref: "#/components/parameters/endTimeRequired"
        - name: limit
          in: query
          required: false
          description: Max traces to export (default 100, max 1000)
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - $ref: "#/components/parameters/sortOrder"
      responses:
        "200":
          description: Export payload
          headers:
            Content-Disposition:
              description: Suggested download filename
              schema:
                type: string
                example: 'attachment; filename="traces-export-20260406-120000.json"'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceExportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ObserverUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/traces/{traceId}/spans/{spanId}:
    get:
      tags:
        - traces
      summary: Get full span detail
      description: |
        Returns the full span including all OTEL attributes and enriched AmpAttributes
        (semantic kind, input/output, token usage, LLM data, tool calls, etc.).
      operationId: getSpanDetail
      security:
        - bearerAuth: []
      parameters:
        - name: traceId
          in: path
          required: true
          description: Trace identifier
          schema:
            type: string
            example: "3cae024cf613a5f37843e9c6eefa3020"
        - name: spanId
          in: path
          required: true
          description: Span identifier
          schema:
            type: string
            example: "a1b2c3d4e5f60001"
      responses:
        "200":
          description: Full enriched span
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichedSpan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ObserverUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Same access token as used for Agent Manager Service API calls.
  parameters:
    organization:
      name: organization
      in: query
      required: true
      description: Organization name
      schema:
        type: string
        example: "default"
    project:
      name: project
      in: query
      required: true
      description: Project name
      schema:
        type: string
    agent:
      name: agent
      in: query
      required: true
      description: Agent name
      schema:
        type: string
    environment:
      name: environment
      in: query
      required: true
      description: Environment name
      schema:
        type: string
...
    startTimeRequired:
      name: startTime
      in: query
      required: true
      schema:
        type: string
        format: date-time
        example: "2026-04-01T00:00:00Z"
    endTimeRequired:
      name: endTime
      in: query
      required: true
      schema:
        type: string
        format: date-time
        example: "2026-04-06T23:59:59Z"
    sortOrder:
      name: sortOrder
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  responses:
    BadRequest:
      description: Bad request — missing or invalid query parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ObserverUnavailable:
      description: Observer Service not configured (OPENCHOREO_OBSERVER_URL unset)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Short error code or literal (e.g. error, Unauthorized)
        message:
          type: string
          description: Human-readable detail
    Span:
      type: object
      required:
        - traceId
        - spanId
        - name
        - service
        - startTime
        - endTime
        - durationInNanos
      properties:
        traceId:
          type: string
        spanId:
          type: string
        parentSpanId:
          type: string
          nullable: true
        name:
          type: string
        service:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationInNanos:
          type: integer
          format: int64
        kind:
          type: string
        status:
          type: string
        attributes:
          type: object
          additionalProperties: true
    FullTrace:
      type: object
      required:
        - traceId
        - rootSpanId
        - rootSpanName
        - startTime
        - endTime
        - spans
      properties:
        traceId:
          type: string
        rootSpanId:
          type: string
        rootSpanName:
          type: string
        rootSpanKind:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationInNanos:
          type: integer
          format: int64
        spanCount:
          type: integer
        tokenUsage:
          $ref: "#/components/schemas/TokenUsage"
        status:
          $ref: "#/components/schemas/TraceStatus"
        input:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        output:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        taskId:
          type: string
          nullable: true
          description: Task ID from baggage (evaluation experiments)
        trialId:
          type: string
          nullable: true
          description: Trial ID from baggage (evaluation experiments)
        spans:
          type: array
          maxItems: 10000
          items:
            $ref: "#/components/schemas/EnrichedSpan"
    TraceExportResponse:
      type: object
      required:
        - traces
        - totalCount
        - truncated
      properties:
        traces:
          type: array
          items:
            $ref: "#/components/schemas/FullTrace"
        totalCount:
          type: integer
        truncated:
          type: boolean
          description: True if any trace hit the MaxSpansPerRequest cap (10 000 spans)

    # ── schemas ─────────────────────────────────────────────────────────────
    TraceOverview:
      type: object
      description: |
        Single trace overview as returned by GET /api/v1/traces.
        Includes root-span-derived input, output, and token usage.
      required:
        - traceId
        - rootSpanId
        - rootSpanName
        - rootSpanKind
        - startTime
        - endTime
        - durationInNanos
        - spanCount
      properties:
        traceId:
          type: string
        rootSpanId:
          type: string
        rootSpanName:
          type: string
        rootSpanKind:
          type: string
          description: Semantic span type (llm, tool, agent, chain, etc.)
          example: llm
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationInNanos:
          type: integer
          format: int64
        spanCount:
          type: integer
        tokenUsage:
          $ref: "#/components/schemas/TokenUsage"
        status:
          $ref: "#/components/schemas/TraceStatus"
        input:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        output:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
    TraceOverviewListResponse:
      type: object
      required:
        - traces
        - totalCount
      properties:
        traces:
          type: array
          items:
            $ref: "#/components/schemas/TraceOverview"
        totalCount:
          type: integer
...
    SpanStatus:
      type: object
      properties:
        error:
          type: boolean
          description: Whether the span has an error
        errorType:
          type: string
          description: Error type from error.type attribute (only present when error is true)
          nullable: true
    AmpAttributes:
      type: object
      description: Semantically enriched attributes added by the AMP platform.
      required:
        - kind
      properties:
        kind:
          type: string
          description: Semantic span type
          enum: [llm, tool, embedding, retriever, agent, crewaitask, chain, unknown]
        input:
          description: Input data (type varies by kind)
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        output:
          description: Output data (type varies by kind)
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        status:
          $ref: "#/components/schemas/SpanStatus"
        data:
          description: Kind-specific data (LLMData, ToolData, EmbeddingData, etc.)
          nullable: true
          type: object
          additionalProperties: true
    EnrichedSpan:
      description: |
        Full span as returned by GET /api/v1/traces/{traceId}/spans/{spanId}.
        Includes raw OTEL attributes and enriched AmpAttributes.
      allOf:
        - $ref: "#/components/schemas/Span"
        - type: object
          properties:
            ampAttributes:
              $ref: "#/components/schemas/AmpAttributes"
    TokenUsage:
      type: object
      properties:
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
    TraceStatus:
      type: object
      properties:
        errorCount:
          type: integer
...
security:
  - bearerAuth: []